Create MACD and EMA Buy & Sell Signals
Using MACD 20, 120 settings and a 5 / 10 ema crossover, this code attempts to proxy what commodity trade advisors might be doing when they buy and sell.  CTAs a systematic traders who trade with the market momentum. 
When the MACD line crosses beneath the signal line (typically after a lengthy bullish run), this signals exhaustion of the move and a correction is likely occuring.  The 5 ema crossing over the 10 ema is added to provide for filtering